+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - scenes can be put in a specified state using the `scm` argument in the `onUpdate` method of scenes, by doing a `scm[scene_id]->setState(State::Idle)
 - adding cross platform shared library loader
 - adding platform identifier
+- adding [imgui_widget_flamegraph](https://github.com/bwrsandman/imgui-flame-graph) and its profiler
 
 ### Changed
 - updating `include/Small/All.hpp` to include the shared library loader